First update of structure
