+++ src/app/Authentification/login/login.component.scss
@@ -1,7 +1,6 @@
 input {
   font-size: 1.3em;
   border: solid lightgray;
-  margin: auto;
 }
 
 .field:focus {
@@ -19,8 +18,7 @@ input {
   width: auto;
   height: 100%;
   display: grid;
-  flex-direction: column;
-  background-image: url(../../assets/light-graybackground.jpg);
+  background-image: url(../../../assets/light-graybackground.jpg);
   background-repeat: no-repeat;
   background-size: cover;
 }
@@ -37,7 +35,6 @@ input {
 .submit {
   width: 275px;
   border-radius: 20px;
-  margin-top: 40px;
   border-color: black;
   background-color: white;
   color: black;
@@ -48,43 +45,70 @@ input {
 }
 .submit:focus {
   color: gold;
-  background-color: black;
 }
-
 .error {
-  border-color: red;
+  border-color: red !important;
 }
-
 .active {
-visibility: visible !important;
-}
+  visibility: visible !important;
+  }
 .invisible{display:none;}
 
-#loginMessage {
-  margin-top: 1em;
-  border: 2px solid gold;
-  background-color: white;
-  height: fit-content;
-  text-align: center;
-  font-size: 1.2em;
-  padding: 1.2em;
-  border-radius: 60px;
-  height: 15em;
-}
-
-#loginForm {
-  text-align: center;
-}
 
 #formContent {
   background-color: white;
   border: 2px solid gold;
   border-radius: 25px;
   display: grid;
-  grid-template-rows: auto auto auto ;
+  grid-template-rows: auto auto auto auto;
   justify-content: center;
   width: 400px;
   height: 530px;
+}
+
+#remMeCont {
+  margin-bottom: 5px;
+}
+#remMe {
+  position: absolute;
+  z-index: -1;
+  opacity: 0;
+}
+#remMe + label {
+  display: inline-flex;
+  align-items: center;
+  user-select: none;
+  font-size: 1.2em;
+}
+#remMe + label::before {
+  content: "";
+  display: inline-block;
+  width: 1em;
+  height: 1em;
+  flex-shrink: 0;
+  flex-grow: 0;
+  border: 1px solid #adb5bd;
+  border-radius: 0.25em;
+  margin-right: 0.5em;
+  background-repeat: no-repeat;
+  background-position: center center;
+  background-size: 50% 50%;
+}
+#remMe:checked + label::before {
+  border-color: black;
+  background-color: gold;
+  background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 8 8'%3e%3cpath fill='000000' d='M6.564.75l-3.59 3.612-1.538-1.55L0 4.26 2.974 7.25 8 2.193z'/%3e%3c/svg%3e");
+}
+
+#errorHandler {
+  height: 40px;
+  margin-bottom: 10px;
+  padding: 5px;
+  visibility: hidden;
+  background-color: white;
+  color: red;
+  text-align: center;
+  font-size: 1.4em;
 }
 
 #inputSec {
@@ -99,29 +123,12 @@ visibility: visible !important;
 }
 
 #password {
-  padding: 5px;
-  border-width: 1px 2px 1px 2px;
-}
-
-#confPassword {
   border-radius: 0px 0px 20px 20px;
   padding: 5px;
   border-width: 1px 2px 2px 2px;
-}
-
-#errorHandler {
-  height: 40px;
-  margin-bottom: 10px;
-  padding: 5px;
-  visibility: hidden;
-  background-color: white;
-  color: red;
-  text-align: center;
-  font-size: 1.4em;
 }
 
 #formFooter {
   justify-content: end;
   padding-bottom: 0px;
 }
-
