+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LoginComponent } from './login/login.component';
-import { RegisterComponent } from './register/register.component';
+import { LoginComponent } from './Authentification/login/login.component';
+import { RegisterComponent } from './Authentification/register/register.component';
 
 const routes: Routes = [
   {path: 'login', component: LoginComponent},
